+++ packages/progress-indicator/src/index.ts
@@ -1,2 +1,3 @@
 export * from './star';
 export * from './dots';
+export * from './circular';