+++ storybook/stories/Components/ProgressIndicator.tsx
@@ -5,6 +5,8 @@ import {
   StarProgressElementProps,
   DotsProgressElement,
   DotsProgressElementProps,
+  CircularProgressElement,
+  CircularProgressElementProps,
 } from '@equinor/fusion-wc-progress-indicator';
 StarProgressElement;
 
@@ -14,4 +16,12 @@ export const StarIndicator = ({ children, ...props }: React.PropsWithChildren<St
 
 export const DotsIndicator = ({ children, ...props }: React.PropsWithChildren<DotsProgressElement>): JSX.Element => {
   return <fwc-dots-progress {...extractProps<DotsProgressElementProps>(props)}>{children}</fwc-dots-progress>;
+
+export const CircularIndicator = ({
+  children,
+  ...props
+}: React.PropsWithChildren<CircularProgressElement>): JSX.Element => {
+  return (
+    <fwc-circular-progress {...extractProps<CircularProgressElementProps>(props)}>{children}</fwc-circular-progress>
+  );
 };