+++ packages/core/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.0.5](https://github.com/equinor/fusion-web-components/compare/@equinor/fusion-wc-core@1.0.4...@equinor/fusion-wc-core@1.0.5) (2022-02-18)
+
+**Note:** Version bump only for package @equinor/fusion-wc-core
+
+
+
+
+
 ## [1.0.4](https://github.com/equinor/fusion-web-components/compare/@equinor/fusion-wc-core@1.0.3...@equinor/fusion-wc-core@1.0.4) (2022-02-08)