+++ packages/list/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.1.22](https://github.com/equinor/fusion-web-components/compare/@equinor/fusion-wc-list@0.1.21...@equinor/fusion-wc-list@0.1.22) (2022-02-18)
+
+**Note:** Version bump only for package @equinor/fusion-wc-list
+
+
+
+
+
 ## [0.1.21](https://github.com/equinor/fusion-web-components/compare/@equinor/fusion-wc-list@0.1.20...@equinor/fusion-wc-list@0.1.21) (2022-02-08)